# use c++ jit dispatch path for op-by-op execution