--- a/jax/api.py
+++ b/jax/api.py
@@ -52,4 +52,3 @@ from jax._src.api import (
   _std_basis,
   _unravel_array_into_pytree,
 )
-