--- a/tests/debug_nans_test.py
+++ b/tests/debug_nans_test.py
@@ -23,6 +23,8 @@ from unittest import SkipTest
 from jax._src import api
 from jax import test_util as jtu
 from jax import numpy as jnp
+from jax._src.lax.lax import div_p
+from jax.interpreters import xla
 from jax.experimental import pjit
 
 from jax.config import config
@@ -53,10 +55,30 @@ class DebugNaNsTest(jtu.JaxTestCase):
     ans.block_until_ready()
 
   def testJitComputationNaN(self):
+    # See the subsequent test as well. This test covers the case where we get a
+    # Python compilation cache hit, since the next line ensures an appropriate
+    # version of the primitive impl rule is compiled. After jaxlib==0.1.65 both
+    # tests likely hit the same code paths, but it doesn't hurt to check both!
+    (0. / jnp.array(1.)).block_until_ready()
     A = jnp.array(0.)
     with self.assertRaises(FloatingPointError):
       ans = jax.jit(lambda x: 0. / x)(A)
       ans.block_until_ready()
+
+  def testJitComputationNaNNoCompilationCache(self):
+    # We make a new primitive so that there's no compilation cache entry for it.
+    new_p = jax.core.Primitive('div')
+    new_p.def_impl(div_p.impl)
+    new_p.def_abstract_eval(div_p.abstract_eval)
+    A = jnp.array(0.)
+
+    xla.translations[new_p] = xla.translations[div_p]
+    try:
+      with self.assertRaises(FloatingPointError):
+        ans = jax.jit(lambda x: new_p.bind(0., x))(A)
+        ans.block_until_ready()
+    finally:
+      del xla.translations[new_p]
 
   def testJitComputationNaNContextManager(self):
     config.update("jax_debug_nans", False)