Historically we had two distinct computation-building and dispatch paths for op-by-op execution (eager mode) and `jit`-decorated functions. Even though the two paths are fundamentally pretty similar (compile an XLA executable or get it from a cache, then execute it on buffers), we specialized the op-by-op path so that it would have lower overheads.

Then much of our `jit` dispatch path got faster (thanks mostly to C++ work by @jblespiau, @hawkinsp, and @zhangqiaorjc). In fact, it got way faster than the op-by-op dispatch path! So we can just unify things under the `jit` compilation and dispatch path. That means simpler code (e.g. only one lowering path, fewer compilation caches (pmap still has its own)), and an op-by-op speed boost too.

Because this change is only to the impl rules of primitives, it's even simpler than the `jit` dispatch path: where the `jit` dispatch path has to worry about Tracer inputs and outputs (and currently bails out to a Python path if Tracers are present), no Tracers on input or output are possible here.

This change mostly consists of
* deleting the separate op-by-op compilation and dispatch paths in xla.py
* replacing them with a C++-API-based dispatch path in xla.py, which looks like a specialized version of some of the `jit` dispatch code in api.py.

The replacement code is not super short, just because setting up a C++ dispatch path is nontrivial. (We could _almost_ achieve what we want in a two-liner with monkey patching...)

~**This change only works with omnistaging.** I marked the change as review-ready because I think we can review it first, even though it's blocked on us deleting the non-omnistaging code paths.~ That's done!

TODO:
- [x] write PR message
- [x] ask hawkinsp@ about debug_nans stuff
- [x] land omnistaging-only
- [ ] ~in a separate PR, add some eager grad benchmarks, compare~

CPU eager benchmarks (just added in #6133):
```
name                   old time/op             new time/op             delta
eager_unary_dispatch   31.6µs ± 2%             21.6µs ± 2%  -31.67%          (p=0.008 n=5+5)
eager_unary            32.3µs ± 1%             22.1µs ± 1%  -31.45%          (p=0.008 n=5+5)
eager_binary_dispatch  39.8µs ± 1%             22.5µs ± 2%  -43.40%          (p=0.008 n=5+5)
eager_binary           40.3µs ± 1%             23.1µs ± 2%  -42.60%          (p=0.008 n=5+5)
```